test: fix test template

+++ gen_float64_seq_test.go
@@ -243,7 +243,7 @@ func TestFloat64Seq_Value(t *testing.T) {
 	data[3].Value = value2
 
 	type args struct {
-		v int64
+		v float64
 	}
 	tests := []struct {
 		name string
